(PC-7905): Move VenueProvider and Provider to core
