+++ src/pcapi/local_providers/install.py
@@ -1,6 +1,6 @@
+from pcapi.core.providers.provider import Provider
 import pcapi.local_providers
 from pcapi.models.db import db
-from pcapi.models.provider import Provider
 from pcapi.repository.provider_queries import get_provider_by_local_class
