+++ src/pcapi/repository/clean_database.py
@@ -2,6 +2,7 @@ from pcapi import settings
 from pcapi.core.mails.models import Email
 from pcapi.core.offerers.models import Offerer
 from pcapi.core.offers.models import Mediation
+from pcapi.core.providers.allocine_venue_provider import AllocineVenueProvider
 from pcapi.core.users.models import Token
 from pcapi.core.users.models import User
 from pcapi.local_providers.install import install_local_providers
@@ -32,7 +33,6 @@ from pcapi.models import Venue
 from pcapi.models import VenueProvider
 from pcapi.models import VenueType
 from pcapi.models.activity import load_activity
-from pcapi.models.allocine_venue_provider import AllocineVenueProvider
 from pcapi.models.db import db
 from pcapi.models.install import install_features
 from pcapi.models.venue_label_sql_entity import VenueLabelSQLEntity
