+++ src/pcapi/routes/serialization/dictifier.py
@@ -4,10 +4,10 @@ from typing import Iterable
 
 from sqlalchemy.orm.collections import InstrumentedList
 
+from pcapi.core.providers.allocine_venue_provider import AllocineVenueProvider
+from pcapi.core.providers.venue_provider import VenueProvider
 from pcapi.domain.reimbursement import BookingReimbursement
-from pcapi.models.allocine_venue_provider import AllocineVenueProvider
 from pcapi.models.pc_object import PcObject
-from pcapi.models.venue_provider import VenueProvider
 from pcapi.routes.serialization.serializer import serialize
