+++ src/pcapi/use_cases/connect_venue_to_provider.py
@@ -1,5 +1,6 @@
 from typing import Optional
 
+from pcapi.core.providers.provider import Provider
 from pcapi.infrastructure.container import api_fnac_stocks
 from pcapi.infrastructure.container import api_libraires_stocks
 from pcapi.infrastructure.container import api_praxiel_stocks
@@ -8,7 +9,6 @@ from pcapi.local_providers.provider_api import synchronize_provider_api
 from pcapi.models import ApiErrors
 from pcapi.models import Venue
 from pcapi.models import VenueProvider
-from pcapi.models.provider import Provider
 from pcapi.repository import repository
